+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * 1.7.3
   - added error message and size for the image comparison
   - fixed sourcePath option for istanbul data
+  - fixed onReceive arguments
 
 * 1.7.2
   - added logging option